backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
openai==0.27.6
python-dotenv==0.19.0
